Fix Gemfile constraint to help proper resolution of tests gems
There is no need to use Criteo's forks of test-kitchen, most patches
have been integrated.
Also it seems that it was not compatible with inspec 2.1.0 and recent
kithen-inspec gem.
We can now remove unecessary version constraint on inspec gem, latest
version is working fine with the rest of the gems.

Remove windows related gems that are not needed:
- chef-zero-scheduled-task
- winrm
- winrm-fs